prevent login/logout from changing module log levels
During init, the shell log backend needs to be enabled because SHELL_START_OBSCURED
prevents it from being initialized.
Enabling the backend requires a param to set the max log priority for all modules.

After init, the backend only needs to be activated or deactivated to prevent log output.

bug 22435